+++ src/slave/constants.hpp
@@ -60,13 +60,13 @@ constexpr Duration DISK_WATCH_INTERVAL = Minutes(1);
 constexpr double GC_DISK_HEADROOM = 0.1;
 
 // Maximum number of completed frameworks to store in memory.
-constexpr uint32_t MAX_COMPLETED_FRAMEWORKS = 50;
+constexpr size_t MAX_COMPLETED_FRAMEWORKS = 50;
 
 // Maximum number of completed executors per framework to store in memory.
-constexpr uint32_t MAX_COMPLETED_EXECUTORS_PER_FRAMEWORK = 150;
+constexpr size_t MAX_COMPLETED_EXECUTORS_PER_FRAMEWORK = 150;
 
 // Maximum number of completed tasks per executor to store in memory.
-constexpr uint32_t MAX_COMPLETED_TASKS_PER_EXECUTOR = 200;
+constexpr size_t MAX_COMPLETED_TASKS_PER_EXECUTOR = 200;
 
 // Default cpus offered by the slave.
 constexpr double DEFAULT_CPUS = 1;
